Broken link
Hi there! I was reading this page and noticed the link to uavcoach.com was broken 404 (the one with the infographic on how drones work).
I tried my best to create a fork / pull request to update the link, but I was unable to figure it out unfortunately.
Anyways, I have written a couple articles that would fit nicely where the old one was!
1. Drone Design - Calculations and assumptions (https://www.tytorobotics.com/blogs/articles/the-drone-design-loop-for-brushless-motors-and-propellers)
2. Types of Drones (https://www.tytorobotics.com/blogs/articles/types-of-drones)
Feel free to use one of them it if you think its a good fit! :)
- Lauren